Made changes to source files in order to properly compile.
Had to move the static functions away from the main nvme_control functions in order to be able to properly use the cmocka mocking functionality.
