+++ spdk/ctest/nvme_control_mocks.c
@@ -21,6 +21,9 @@
 // portions thereof marked with this legend must also reproduce the markings.
 //
 
+#include <stdarg.h>
+#include <stddef.h>
+#include <setjmp.h>
 #include <cmocka.h>
 
 #include "spdk/stdinc.h"
@@ -28,7 +31,19 @@
 #include "spdk/env.h"
 
 #include "nvme_control.h"
-#include "nvme_control_mocks.h"
+
+struct ctrlr_entry {
+	struct spdk_nvme_ctrlr	*ctrlr;
+	const char		*tr_addr;
+	struct ctrlr_entry	*next;
+};
+
+struct ns_entry {
+	struct spdk_nvme_ctrlr	*ctrlr;
+	struct spdk_nvme_ns	*ns;
+	struct ns_entry		*next;
+	struct spdk_nvme_qpair	*qpair;
+};
 
 /**
  * ==============================
@@ -38,17 +53,20 @@
 
 struct ret_t * __wrap_init_ret(void)
 {
-	return mock_type(ret_t *);
+	printf("Mock init_ret() runnning...");
+	return mock_ptr_type(struct ret_t *);
 }
 
 void __wrap_cleanup(void)
 {
+	printf("Mock cleanup() running...");
 	struct ns_entry		*ns_entry;
 	struct ctrlr_entry	*ctrlr_entry;
 
-	ns_entry = mock_type(ns_entry *);
-	ctrlr_entry = mock_type(ctrlr_entry);
+	//ns_entry = g_namespaces;
+	//ctrlr_entry = g_controllers;
 
+	/*
 	while (ns_entry) {
 		struct ns_entry *next = ns_entry->next;
 
@@ -62,11 +80,114 @@ void __wrap_cleanup(void)
 		test_free(ctrlr_entry);
 		ctrlr_entry = next;
 	}
+	*/
 }
 
 void __wrap_collect(struct ret_t *ret)
 {
-	check_expected_ptr(ret);
+	printf("Mock collect() running...");
+	struct ns_entry				*ns_entry;
+	struct ctrlr_entry			*ctrlr_entry;
+	const struct spdk_nvme_ctrlr_data	*cdata;
+	int					written;
+
+	//ns_entry = g_namespaces;
+	//ctrlr_entry = g_controllers;
+
+	/*
+	while (ns_entry) {
+		struct ns_t *ns_tmp;
+
+		ns_tmp = test_malloc(sizeof(struct ns_t));
+
+		if (ns_tmp == NULL) {
+			snprintf(ret->err, sizeof(ret->err), "ns_t malloc");
+			ret->rc = -ENOMEM;
+			return;
+		}
+
+		cdata = spdk_nvme_ctrlr_get_data(ns_entry->ctrlr);
+
+		ns_tmp->id = spdk_nvme_ns_get_id(ns_entry->ns);
+		// capacity in GBytes
+		ns_tmp->size = spdk_nvme_ns_get_size(ns_entry->ns) / \
+			       NVMECONTROL_GBYTE_BYTES;
+
+		if (set_pci_addr(
+			ns_entry->ctrlr, ns_tmp->ctrlr_pci_addr,
+			sizeof(ns_tmp->ctrlr_pci_addr), ret) != 0) {
+
+			return;
+		}
+
+		ns_tmp->next = ret->nss;
+		ret->nss = ns_tmp;
+
+		ns_entry = ns_entry->next;
+	}
+
+	while (ctrlr_entry) {
+		struct ctrlr_t *ctrlr_tmp;
+
+		ctrlr_tmp = test_malloc(sizeof(struct ctrlr_t));
+
+		if (ctrlr_tmp == NULL) {
+			perror("ctrlr_t malloc");
+			ret->rc = -ENOMEM;
+			return;
+		}
+
+		cdata = spdk_nvme_ctrlr_get_data(ctrlr_entry->ctrlr);
+
+		written = snprintf(
+			ctrlr_tmp->model, sizeof(ctrlr_tmp->model),
+			"%-20.20s", cdata->mn
+		);
+		if (check_size(
+			written, sizeof(ctrlr_tmp->model),
+			"model truncated", ret) != 0) {
+
+			return;
+		}
+
+		written = snprintf(
+			ctrlr_tmp->serial, sizeof(ctrlr_tmp->serial),
+			"%-20.20s", cdata->sn
+		);
+		if (check_size(
+			written, sizeof(ctrlr_tmp->serial),
+			"serial truncated", ret) != 0) {
+
+			return;
+		}
+
+		written = snprintf(
+			ctrlr_tmp->fw_rev, sizeof(ctrlr_tmp->fw_rev),
+			"%s", cdata->fr
+		);
+		if (check_size(
+			written, sizeof(ctrlr_tmp->fw_rev),
+			"firmware revision truncated", ret) != 0) {
+
+			return;
+		}
+
+		if (set_pci_addr(
+			ctrlr_entry->ctrlr, ctrlr_tmp->pci_addr,
+			sizeof(ctrlr_tmp->pci_addr), ret) != 0) {
+
+			return;
+		}
+
+		// cdata->cntlid is not unique per host, only per subsystem
+		ctrlr_tmp->next = ret->ctrlrs;
+		ret->ctrlrs = ctrlr_tmp;
+
+		ctrlr_entry = ctrlr_entry->next;
+	}
+
+	ret->rc = 0;
+	*/
 }
 
 /**
@@ -82,12 +203,12 @@ int __wrap_spdk_nvme_probe(
 		spdk_nvme_attach_cb attach_cb, 
 		spdk_nvme_remove_cb remove_cb)
 {
-	/* Check parameters */
-	check_expected_ptr(trid);
-	check_expected_ptr(cb_trix);
-	check_expected(probe_cb);
-	check_expected(attach_cb);
-	check_expected(remove_cb);
+	printf("Mock spdk_nvme_probe() running...");
 
+	/* Check parameters */
+	//check_expected(probe_cb);
+	//check_expected(attach_cb);
+
+	/* Return integer */
 	return mock_type(int);
 }
